Changed search index update script to use a Monitor instead of a CoverageProvider.
